+++ packages/@runlightyear/lightyear/src/util/isFunction.ts
@@ -1,3 +1,5 @@
-export default function isFunction(val: unknown) {
+export function isFunction(val: unknown) {
   return val instanceof Function;
 }
+
+export default isFunction;