+++ packages/@runlightyear/postmark/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @runlightyear/postmark
 
+## 0.1.9
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies
+  - @runlightyear/lightyear@0.23.0
+
 ## 0.1.8
 
 ### Patch Changes