+++ packages/@runlightyear/salesforce/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @runlightyear/salesforce
 
+## 0.2.2
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies
+  - @runlightyear/lightyear@0.23.0
+
 ## 0.2.1
 
 ### Patch Changes